인솔(인솔)/workspace/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
<<<<<<< Updated upstream:backend/requirements.txt
websockets==15.0.1
python-socketio==5.13.0
google-generativeai>=0.8.5 
=======
requests==2.31.0
google-generativeai==0.3.2
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
openai>=1.0.0
# HuggingFace 모델 임시 비활성화 (빌드 시간 단축)
# sentence-transformers==2.2.2
# scikit-learn==1.3.2
# numpy==1.24.3 
>>>>>>> Stashed changes:admin/backend/requirements.txt
